Fix integer overflow in pipeline coordinate calculations
- Cast screen coordinates to `i64` before subtraction in `fill_triangle_3d` and `fill_triangle_gouraud` to prevent `attempt to subtract with overflow` panics when coordinates are extreme (e.g. near-singular projection).
- Use `-(x as f32)` pattern for negating screen coordinates to prevent `attempt to negate with overflow` when `x` is `i32::MIN`.
- Add `tests/pipeline_overflow.rs` to reproduce the issue and prevent regression.
- Update Sentry journal with findings.

+++ src/pipeline.rs
@@ -125,7 +125,7 @@ fn draw_scanline_flat(
 
     if xs < 0 {
         // Advance z if we start off-screen
-        z += (-xs) as f32 * dz_dx;
+        z += -(xs as f32) * dz_dx;
         xs = 0;
     }
 
@@ -191,17 +191,17 @@ pub fn fill_triangle_3d(
             continue;
         };
 
-        let mut ax = p0.x as f32 + (p2.x - p0.x) as f32 * step.alpha;
+        let mut ax = p0.x as f32 + (p2.x as i64 - p0.x as i64) as f32 * step.alpha;
         let mut az = p0.z + (p2.z - p0.z) * step.alpha;
 
         let (mut bx, mut bz) = if step.second_half {
             (
-                p1.x as f32 + (p2.x - p1.x) as f32 * step.beta,
+                p1.x as f32 + (p2.x as i64 - p1.x as i64) as f32 * step.beta,
                 p1.z + (p2.z - p1.z) * step.beta,
             )
         } else {
             (
-                p0.x as f32 + (p1.x - p0.x) as f32 * step.beta,
+                p0.x as f32 + (p1.x as i64 - p0.x as i64) as f32 * step.beta,
                 p0.z + (p1.z - p0.z) * step.beta,
             )
         };
@@ -214,7 +214,7 @@ pub fn fill_triangle_3d(
         let x_start = ax as i32;
         let x_end = bx as i32;
 
-        let dx = x_end - x_start;
+        let dx = x_end as i64 - x_start as i64;
         let dz = bz - az;
 
         // Handle single pixel or invalid width
@@ -347,7 +347,7 @@ pub fn fill_triangle_gouraud(
 
     // Calculate gradients for the long edge (p0 -> p2)
     let inv_total_height = 1.0 / total_height;
-    let dx_dy_a = (p2.x - p0.x) as f32 * inv_total_height;
+    let dx_dy_a = (p2.x as i64 - p0.x as i64) as f32 * inv_total_height;
     let dz_dy_a = (p2.z - p0.z) * inv_total_height;
     let dc_dy_a = (c2 - c0) * inv_total_height;
 
@@ -356,7 +356,7 @@ pub fn fill_triangle_gouraud(
     // x_long = p0.x + (p2.x - p0.x) * (p1.y - p0.y) / (p2.y - p0.y)
     let dy_total = p2.y - p0.y;
     let x_long_at_p1 = if dy_total != 0 {
-        p0.x as f32 + (p2.x - p0.x) as f32 * ((p1.y - p0.y) as f32 / dy_total as f32)
+        p0.x as f32 + (p2.x as i64 - p0.x as i64) as f32 * ((p1.y - p0.y) as f32 / dy_total as f32)
     } else {
         p0.x as f32
     };
@@ -378,7 +378,7 @@ pub fn fill_triangle_gouraud(
     let (dx_dy_b1, dz_dy_b1, dc_dy_b1) = if h1 != 0.0 {
         let inv_h1 = 1.0 / h1;
         (
-            (p1.x - p0.x) as f32 * inv_h1,
+            (p1.x as i64 - p0.x as i64) as f32 * inv_h1,
             (p1.z - p0.z) * inv_h1,
             (c1 - c0) * inv_h1,
         )
@@ -407,7 +407,7 @@ pub fn fill_triangle_gouraud(
             let h2 = (p2.y - p1.y) as f32;
             if h2 != 0.0 {
                 let inv_h2 = 1.0 / h2;
-                let dx_dy_b2 = (p2.x - p1.x) as f32 * inv_h2;
+                let dx_dy_b2 = (p2.x as i64 - p1.x as i64) as f32 * inv_h2;
                 let dz_dy_b2 = (p2.z - p1.z) * inv_h2;
                 let dc_dy_b2 = (c2 - c1) * inv_h2;
 
@@ -429,7 +429,7 @@ pub fn fill_triangle_gouraud(
         let h2 = (p2.y - p1.y) as f32;
         if h2 != 0.0 {
             let inv_h2 = 1.0 / h2;
-            dx_dy_b = (p2.x - p1.x) as f32 * inv_h2;
+            dx_dy_b = (p2.x as i64 - p1.x as i64) as f32 * inv_h2;
             dz_dy_b = (p2.z - p1.z) * inv_h2;
             dc_dy_b = (c2 - c1) * inv_h2;
         }
@@ -445,7 +445,7 @@ pub fn fill_triangle_gouraud(
             let h2 = (p2.y - p1.y) as f32;
             if h2 != 0.0 {
                 let inv_h2 = 1.0 / h2;
-                dx_dy_b = (p2.x - p1.x) as f32 * inv_h2;
+                dx_dy_b = (p2.x as i64 - p1.x as i64) as f32 * inv_h2;
                 dz_dy_b = (p2.z - p1.z) * inv_h2;
                 dc_dy_b = (c2 - c1) * inv_h2;
             }
@@ -460,7 +460,7 @@ pub fn fill_triangle_gouraud(
 
         let x_start = x_left as i32;
         let x_end = x_right as i32;
-        let dx = x_end - x_start;
+        let dx = x_end as i64 - x_start as i64;
 
         if dx <= 0 {
             if x_start >= 0 && x_start < width as i32 && zb.test_and_set(x_start, y, z_left) {
@@ -490,7 +490,7 @@ pub fn fill_triangle_gouraud(
 
         // Clamp to screen bounds
         if xs < 0 {
-            let diff = -xs as f32;
+            let diff = -(xs as f32);
             z += diff * dz_dx;
             c = c + dc_dx * diff;
             xs = 0;
